@@ -2,17 +2,44 @@ plugins {
     id 'com.android.application'
 }
 
+import java.util.Properties
+
+// --- Secrets loader: apikey.properties (or environment variables as fallback) ---
+def secrets = new Properties()
+def secretsFile = rootProject.file("apikey.properties")
+if (secretsFile.exists()) {
+    secrets.load(new FileInputStream(secretsFile))
+}
+def readSecret = { String key, String defVal = "" ->
+    def fromEnv = System.getenv(key)
+    return (fromEnv != null && !fromEnv.isEmpty()) ? fromEnv : (secrets.getProperty(key, defVal) ?: defVal)
+}
+def DEFAULT_BACK4APP_URL = "https://parseapi.back4app.com/"
+// -------------------------------------------------------------------------------
+
 android {
-    compileSdk 32
+    namespace 'com.example.fludde'
+    compileSdk 35
 
     defaultConfig {
         applicationId "com.example.fludde"
         minSdk 21
-        targetSdk 32
+        targetSdk 35
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        vectorDrawables {
+            useSupportLibrary true
+        }
+
+        buildConfigField "String", "SPOTIFY_KEY", "\"${readSecret('SPOTIFY_KEY')}\""
+        buildConfigField "String", "TMDB_API_KEY", "\"${readSecret('TMDB_API_KEY')}\""
+        buildConfigField "String", "NY_TIMES_API_KEY", "\"${readSecret('NY_TIMES_API_KEY')}\""
+
+        // Back4App / Parse
+        buildConfigField "String", "BACK4APP_APP_ID", "\"${readSecret('BACK4APP_APP_ID')}\""
+        buildConfigField "String", "BACK4APP_CLIENT_KEY", "\"${readSecret('BACK4APP_CLIENT_KEY')}\""
+        buildConfigField "String", "BACK4APP_SERVER_URL", "\"${readSecret('BACK4APP_SERVER_URL', DEFAULT_BACK4APP_URL)}\""
     }
 
     buildTypes {
