@@ -1,6 +1,7 @@
 package com.example.fludde.fragments;
 
 import android.annotation.SuppressLint;
+import android.content.Context;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -9,83 +10,61 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.text.TextUtils;
 import android.util.Log;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.inputmethod.EditorInfo;
+import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.Toast;
 
 import com.example.fludde.R;
-import com.example.fludde.User;
 import com.example.fludde.adapters.SearchFragmentAdapter;
-import com.parse.FindCallback;
+import com.google.android.material.textfield.TextInputEditText;
 import com.parse.ParseException;
 import com.parse.ParseQuery;
 import com.parse.ParseUser;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 
 /**
- * A simple {@link Fragment} subclass.
- * Use the {@link SearchFragment#newInstance} factory method to
- * create an instance of this fragment.
+ * Search screen:
+ * - Prominent search field + primary button (consistent placement)
+ * - Empty state guidance when no results
+ * - Keyboard action (IME_ACTION_SEARCH / Enter) equals tapping the button
  */
 public class SearchFragment extends Fragment {
 
-    List<ParseUser> allUsers;
+    private static final String TAG = "SearchFragment";
+
     private RecyclerView rvUserQuery;
-    SearchFragmentAdapter searchFragmentAdapter;
-    EditText etSearchFieldUser;
-    Button btSearch;
+    private SearchFragmentAdapter searchFragmentAdapter;
+    private final List<ParseUser> allUsers = new ArrayList<>();
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    private TextInputEditText etSearchFieldUser;
+    private Button btSearch;
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
+    // Empty state views
+    private View emptyState;
 
-    public SearchFragment() {
-        // Required empty public constructor
-    }
+    public SearchFragment() { }
 
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment SearchFragment.
-     */
-    // TODO: Rename and change types and number of parameters
     public static SearchFragment newInstance(String param1, String param2) {
         SearchFragment fragment = new SearchFragment();
         Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
+        args.putString("param1", param1);
+        args.putString("param2", param2);
         fragment.setArguments(args);
         return fragment;
     }
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
-
-    @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_search, container, false);
     }
 
@@ -96,65 +75,93 @@ public class SearchFragment extends Fragment {
         rvUserQuery = view.findViewById(R.id.rvUserQuery);
         etSearchFieldUser = view.findViewById(R.id.etSearchFieldUser);
         btSearch = view.findViewById(R.id.btSearch);
+        emptyState = view.findViewById(R.id.emptyState);
 
-        allUsers = new ArrayList<>();
-        searchFragmentAdapter = new SearchFragmentAdapter(getContext(), allUsers);
-
+        searchFragmentAdapter = new SearchFragmentAdapter(requireContext(), allUsers);
         rvUserQuery.setAdapter(searchFragmentAdapter);
         rvUserQuery.setLayoutManager(new LinearLayoutManager(getContext()));
 
-        btSearch.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
+        // Primary button tap
+        btSearch.setOnClickListener(v -> performSearch(getQueryText()));
 
-                String userName = etSearchFieldUser.getText().toString();
-
-                lookForUser(userName);
+        // Keyboard action (Search / Enter) == button tap
+        etSearchFieldUser.setOnEditorActionListener((tv, actionId, event) -> {
+            boolean handled = false;
+            if (actionId == EditorInfo.IME_ACTION_SEARCH) {
+                performSearch(getQueryText());
+                handled = true;
+            } else if (event != null
+                    && event.getKeyCode() == KeyEvent.KEYCODE_ENTER
+                    && event.getAction() == KeyEvent.ACTION_DOWN) {
+                performSearch(getQueryText());
+                handled = true;
             }
+            return handled;
         });
 
+        // Show initial guidance
+        showEmptyState(true);
     }
 
-//    private void lookForUser(String userName) {
-//        ParseQuery<User> query =  ParseQuery.getQuery(User.class);
-//        query.whereEqualTo(User.KEY_USER, userName.toString());
-//        query.findInBackground(new FindCallback<User>() {
-//            @Override
-//            public void done(List<User> users, ParseException e) {
-//                if (e == null) {
-//                    // The query was successful, returns the users that matches
-//                    // the criteria.
-//                    for(User aUser : users) {
-//                        aUser.getUsername().toString();
-//                    }
-//                } else {
-//                    // Something went wrong.
-//                Toast.makeText(getContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
-//                }
-//                allUsers.addAll(users);
-//                searchFragmentAdapter.notifyDataSetChanged();
-//
-//            }
-//        });
+    private String getQueryText() {
+        String q = etSearchFieldUser.getText() != null ? etSearchFieldUser.getText().toString().trim() : "";
+        return q;
+    }
 
+    private void performSearch(String userName) {
+        hideKeyboard();
+        if (TextUtils.isEmpty(userName)) {
+            Toast.makeText(getContext(), getString(R.string.hint_search_user), Toast.LENGTH_SHORT).show();
+            allUsers.clear();
+            searchFragmentAdapter.notifyDataSetChanged();
+            showEmptyState(true);
+            return;
+        }
+        lookForUser(userName);
+    }
+
+    private void showEmptyState(boolean show) {
+        if (emptyState != null) emptyState.setVisibility(show ? View.VISIBLE : View.GONE);
+    }
+
+    private void hideKeyboard() {
+        try {
+            InputMethodManager imm = (InputMethodManager) requireContext().getSystemService(Context.INPUT_METHOD_SERVICE);
+            View v = requireActivity().getCurrentFocus();
+            if (imm != null && v != null) {
+                imm.hideSoftInputFromWindow(v.getWindowToken(), 0);
+            }
+        } catch (Exception ignore) { }
+    }
 
     private void lookForUser(String userName) {
+        // Make the same action work for button and keyboard
         ParseQuery<ParseUser> query = ParseUser.getQuery();
-        query.whereEqualTo("username", userName);
+
+        // Use a contains match for a friendlier search experience
+        query.whereContains("username", userName);
+
+        // Optional: small limit to keep UI snappy; adjust as needed
+        query.setLimit(50);
+
         query.findInBackground((users, e) -> {
-            if (e == null) {
-                // The query was successful, returns the users that matches
-                // the criteria.
-                for(ParseUser aUser : users) {
-                    aUser.getUsername().toString();
-                }
-            } else {
-                // Something went wrong.
-
+            if (e != null) {
+                Log.e(TAG, "Search failed", e);
+                Toast.makeText(getContext(), getString(R.string.error_search), Toast.LENGTH_SHORT).show();
+                allUsers.clear();
+                searchFragmentAdapter.notifyDataSetChanged();
+                showEmptyState(true);
+                return;
             }
-            allUsers.addAll(users);
-          searchFragmentAdapter.notifyDataSetChanged();
-        });;
-    }
 
+            allUsers.clear();
+            if (users != null) {
+                allUsers.addAll(users);
+            }
+            searchFragmentAdapter.notifyDataSetChanged();
+
+            // Empty state guidance if nothing found
+            showEmptyState(allUsers.isEmpty());
+        });
+    }
 }
