@@ -7,17 +7,16 @@ import com.parse.ParseObject;
 public class ParseApplication extends Application {
     @Override
     public void onCreate() {
-
         super.onCreate();
 
         ParseObject.registerSubclass(Post.class);
         ParseObject.registerSubclass(User.class);
 
+        // Read secrets from BuildConfig (injected by Gradle)
         Parse.initialize(new Parse.Configuration.Builder(this)
-                .applicationId(getString(R.string.back4app_app_id))
-                .clientKey(getString(R.string.back4app_client_key))
-                .server(getString(R.string.back4app_server_url))
+                .applicationId(BuildConfig.BACK4APP_APP_ID)
+                .clientKey(BuildConfig.BACK4APP_CLIENT_KEY)
+                .server(BuildConfig.BACK4APP_SERVER_URL)
                 .build());
     }
-    }
-
+}
