@@ -13,10 +13,13 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.TextView;
 
 import com.example.fludde.Post;
 import com.example.fludde.adapters.PostAdapter;
 import com.example.fludde.R;
+import com.facebook.shimmer.ShimmerFrameLayout;
 import com.parse.FindCallback;
 import com.parse.ParseException;
 import com.parse.ParseQuery;
@@ -24,62 +27,22 @@ import com.parse.ParseQuery;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link PostFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
 public class PostFragment extends Fragment {
-     private RecyclerView rvPost;
+    private RecyclerView rvPost;
     private static final String TAG = "Post";
-     protected PostAdapter adapter;
+    protected PostAdapter adapter;
+    protected List<Post> allPosts;
 
-     protected List<Post> allPosts;
+    private ShimmerFrameLayout shimmer;
+    private View errorCard;
+    private TextView tvErrorMessage;
+    private Button btnRetry;
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-    public PostFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment PostFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static PostFragment newInstance(String param1, String param2) {
-        PostFragment fragment = new PostFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
+    public PostFragment() {}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_post, container, false);
     }
 
@@ -87,18 +50,49 @@ public class PostFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         rvPost = view.findViewById(R.id.rvPost);
-        allPosts = new ArrayList<>();
+        shimmer = view.findViewById(R.id.shimmerContainer);
 
+        errorCard = view.findViewById(R.id.inlineError);
+        tvErrorMessage = view.findViewById(R.id.tvErrorMessage);
+        btnRetry = view.findViewById(R.id.btnRetry);
+
+        allPosts = new ArrayList<>();
         adapter = new PostAdapter(getContext(), allPosts);
 
-
-        ///
         rvPost.setAdapter(adapter);
         rvPost.setLayoutManager(new LinearLayoutManager(getContext()));
-        queryPosts();
+        rvPost.setHasFixedSize(true);
 
+        btnRetry.setOnClickListener(v -> queryPosts());
+
+        showError(false, null);
+        showLoading(true);
+        queryPosts();
     }
+
+    private void showLoading(boolean show) {
+        if (show) {
+            shimmer.setVisibility(View.VISIBLE);
+            shimmer.startShimmer();
+            rvPost.setAlpha(0f);
+        } else {
+            shimmer.stopShimmer();
+            shimmer.setVisibility(View.GONE);
+            rvPost.setAlpha(1f);
+        }
+    }
+
+    private void showError(boolean show, @Nullable String message) {
+        errorCard.setVisibility(show ? View.VISIBLE : View.GONE);
+        if (show) {
+            tvErrorMessage.setText(message != null ? message : getString(R.string.error_load_feed));
+        }
+    }
+
     protected void queryPosts() {
+        showError(false, null);
+        showLoading(true);
+
         ParseQuery<Post> query = ParseQuery.getQuery(Post.class);
         query.include(Post.KEY_USER);
         query.setLimit(20);
@@ -107,14 +101,13 @@ public class PostFragment extends Fragment {
             @SuppressLint("NotifyDataSetChanged")
             @Override
             public void done(List<Post> posts, ParseException e) {
-                if (e!=null)
-                {
-                    Log.e("post", "post error");
+                showLoading(false);
+                if (e != null) {
+                    Log.e(TAG, "post error", e);
+                    showError(true, getString(R.string.error_load_feed));
+                    return;
                 }
-                for (Post post : posts){
-                    Log.i("Post Query", "Post" +
-                            post.getDescription());
-                }
+                allPosts.clear();
                 allPosts.addAll(posts);
                 adapter.notifyDataSetChanged();
             }
